move aggregate into main
